updating fbx2ndMesh toll (wip)

+++ newton-4.00/applications/toolsAndWrapers/fbxToMesh/install/source/ndUtils.h
@@ -209,7 +209,7 @@ inline void ndMemSet(T* const dst, const T& val, ndInt64 elements)
 template <class T>
 inline void ndMemCpy(T* const dst, const T* const src, ndInt64 elements)
 {
-	size_t bytes = sizeof(T) * elements;
+	size_t bytes = sizeof(T) * size_t(elements);
 	memcpy((void*)dst, (void*)src, bytes);
 }
